Refactor startup dialogs
